requirements.txt
astroid==2.1.0
boto3==1.9.101
botocore==1.12.101
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==1.11.20
django-cors-headers==2.4.0
django-storages==1.7.1
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
docutils==0.14
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
mysqlclient==2.0.3
Pillow==5.4.1
pkg-resources==0.0.0
psycopg2-binary==2.7.7
PyJWT==1.7.1
pylint==2.2.2
python-dateutil==2.8.0
python-decouple==3.1
python-dotenv==0.15.0
pytz==2018.9
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.11.1
